+++ functions/lib/modules/payments/endpoints.d.ts
@@ -1,8 +1,0 @@
-import * as functions from 'firebase-functions';
-export declare const createPaymentEndpoint: functions.HttpsFunction;
-export declare const paymentWebhookEndpoint: functions.HttpsFunction;
-export declare const getPlansEndpoint: functions.HttpsFunction;
-export declare const validatePromoEndpoint: functions.HttpsFunction;
-export declare const refundEndpoint: functions.HttpsFunction;
-export declare const paymentSuccessEndpoint: functions.HttpsFunction;
-//# sourceMappingURL=endpoints.d.ts.map