+++ functions/lib/modules/sync/index.d.ts
@@ -1,4 +1,0 @@
-export { storeApiKey, getApiKey, deleteApiKey, hasApiKey } from './secretManager';
-export { fetchWBCards, fetchWBPrices, zeroWBStock, updateWBPrice, mapWBCardToProduct } from './wbFetcher';
-export { fetchOzonProducts, fetchOzonProductInfo, fetchOzonPrices, zeroOzonStock, updateOzonPrice, mapOzonProductToProduct } from './ozonFetcher';
-//# sourceMappingURL=index.d.ts.map