+++ functions/src/modules/sync/index.ts
@@ -1,3 +1,0 @@
-export { storeApiKey, getApiKey, deleteApiKey, hasApiKey } from './secretManager';
-export { fetchWBCards, fetchWBPrices, zeroWBStock, updateWBPrice, mapWBCardToProduct } from './wbFetcher';
-export { fetchOzonProducts, fetchOzonProductInfo, fetchOzonPrices, zeroOzonStock, updateOzonPrice, mapOzonProductToProduct } from './ozonFetcher';